+++ compiler-tests/src/test/data/box/inject/member/ZeroArgumentMethodInjection.kt
@@ -1,0 +1,20 @@
+class Thing {
+  var initialized = false
+
+  @Inject
+  fun onInit() {
+    initialized = true
+  }
+}
+
+@DependencyGraph
+interface AppGraph {
+  val thingInjector: MembersInjector<Thing>
+}
+
+fun box(): String {
+  val thing = Thing()
+  createGraph<AppGraph>().thingInjector.injectMembers(thing)
+  assertTrue(thing.initialized)
+  return "OK"
+}